==> 02_Spacial Join.ipynb <==
#### Libraries

import geopandas as gpd
import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings('ignore')

#### Importing Data

#Ecoregion data from EPA
ecoregions = gpd.read_file("us_eco_l3.shp")

#MTBS Wildfire data
mtbs_fires = gpd.read_file("mtbs_fod_pts_DD.shp")

#Climate Monitor Locations
metadata_col_specs = [(0,12),(12,21),(21,31),(31,38),(38,41),(41,72),(72,76),(76,80),(80,86)]

metadata_names = ["ID","LATITUDE","LONGITUDE","ELEVATION","STATE","NAME","GSN FLAG","HCN/CRN FLAG","WMO ID"]

metadata_dtype = {"ID": str, "LATITUDE": float, "LONGITUDE":float, "ELEVATION": float, "STATE": str,"NAME": str,
    "GSN FLAG": str,"HCN/CRN FLAG": str,"WMO ID": str}

def read_station_metadata(filename="ghcnd-stations.txt"):
    """Reads in station metadata

    :filename: ghcnd station metadata file.
    :returns: station metadata as a pandas Dataframe

    """
    df = pd.read_fwf(filename, metadata_col_specs, names=metadata_names, dtype=metadata_dtype)

    return df

monitors = read_station_metadata()

#### Checking/Cleaning Data

#Selecting columns from ecoregion data
eco_join = ecoregions[['NA_L3NAME', 'NA_L2NAME', 'NA_L1NAME', 'geometry']]

#Converting monitor dataframe to geodataframe
monitor_geo = gpd.GeoDataFrame(
    monitors, geometry=gpd.points_from_xy(monitors.LONGITUDE, monitors.LATITUDE))
monitor_geo = monitor_geo.set_crs('EPSG:4269')

#Checking to make sure Coordinate Reference Systems match
print(eco_join.crs == mtbs_fires.crs)
print(monitor_geo.crs == mtbs_fires.crs)

#Matching the CRS of both Ecoregions and Monitors to the MTBS database
eco_join = eco_join.to_crs(mtbs_fires.crs)
monitor_geo = monitor_geo.to_crs(mtbs_fires.crs)

print(eco_join.crs == mtbs_fires.crs)
print(monitor_geo.crs == mtbs_fires.crs)

#### Joining Data

#Join MTBS and Ecoregions
eco_mtbs = gpd.sjoin(mtbs_fires, eco_join, how='left', op='intersects')
eco_mtbs.drop(columns=['geometry', 'index_right'], inplace=True)

#Join Monitors and Ecoregions
eco_monitor = gpd.sjoin(monitor_geo, eco_join, how='left', op='intersects')

#### Sorting Joined Data

#US only weather monitors
eco_monitor_us = eco_monitor[eco_monitor['ID'].str.contains('US')]
eco_monitor_us.drop(columns=['NAME', 'GSN FLAG', 'HCN/CRN FLAG', 'WMO ID', 'geometry', 'index_right'],
                 inplace=True)
eco_monitor_us.head()

#creating a monitor database that have only monitors in ecoregions where fires have occured
sub_regions_mtbs = eco_mtbs['NA_L3NAME'].unique()
eco_monitor_mtbs = eco_monitor_us[eco_monitor_us['NA_L3NAME'].isin(sub_regions_mtbs)]

#### Writing Data to CSVs

eco_mtbs.to_csv('MTBS_and_Region.csv')

eco_monitor_mtbs.to_csv('Monitor_by_Region.csv')
