==> 06_Non-parametric Modeling.ipynb <==
## Introduction

This will be one last attempt at trying to create a model for this dataset. As I've already explained, there is still quite a bit of information missing for a comprehensive model, but I'm going to attempt to use a non-parametric algorithm. Up until this point, I had just been using different iterations of a linear regression model, which the data has proved does not fit that trend.

#### Libraries

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import train_test_split
from sklearn.tree import export_graphviz
import pydot
from sklearn import tree
import graphviz

import warnings
warnings.filterwarnings('ignore')

#### Data

#MTBS Data
mtbs = pd.read_csv('MTBS_clean.csv')
mtbs.drop(columns='Unnamed: 0', inplace=True)

#Ecoregion Level 3 - Weather Data
weather = pd.read_csv('Ecoregion Level 3 Weather.csv', skiprows=[1])
weather.drop(columns='Unnamed: 0', inplace=True)

#### Processing

#Converting to date
weather['Date'] = pd.to_datetime(weather['Date'], format='%Y-%m-%d')
mtbs['Date'] = pd.to_datetime(mtbs['Date'], format='%Y-%m-%d')

#Joining Data
mtbs_weather_3 = pd.merge(mtbs, weather, how='left', on=['NA_L3NAME', 'Date'])
mtbs_weather_3.drop(columns=['wind_speed_ms_30', 'wind_speed_ms_90', 'wind_speed_ms_180', 'wind_speed_ms_365'],
                   inplace=True)
mtbs_weather_3.dropna(inplace=True)

## Modeling

col_keep = ['Fire_Type', 'Low_T', 'Mod_T', 'High_T', 'NA_L3NAME', 'state', 'month_ig', 'max_temp_C_90',
           'precipitation_cm_90', 'Acres']

df_model_1 = mtbs_weather_3[col_keep]
df_model_1 = df_model_1[df_model_1['Fire_Type'] == 'Wildfire']
target_1 = np.array(df_model_1['Acres'])
predictors_1 = df_model_1.drop(columns=['Acres'])
predictors_1 = pd.get_dummies(predictors_1, drop_first = True)

X_train, X_test, y_train, y_test = train_test_split(predictors_1, target_1, test_size=0.2)

# Instantiate model with 1000 decision trees
rf = RandomForestRegressor(n_estimators = 500, random_state = 42)
# Train the model on training data
rf.fit(X_train, y_train)

rf.score(X_train, y_train)

rf.get_params()

predictions = rf.predict(X_test)
errors = abs(predictions - y_test)
print('Mean Absolute Error:', round(np.mean(errors), 2), 'acres.')

importances = list(rf.feature_importances_)
feature_importances = [(feature, round(importance, 2)) for feature, importance in zip(X_train.columns, importances)]
feature_importances = sorted(feature_importances, key = lambda x: x[1], reverse = True)
[print('Variable: {:20} Importance: {}'.format(*pair)) for pair in feature_importances]

#### Conclusion

Well, the Random Forest Regressor was a mixed bag. The r-squared value from the RFR model was much better than any of the r-squared values returned in the linear regression models, but my accuracy and mean absolute error didn't look that good. Below are the performance metrics for my model:
1. R-squared: `0.855`
2. Mean absolute error: `10,866`

This is a classic example of an overfit model. My regressor trained really well on my training data, but did a bad job predicting the burn size of my test data.
